Extract user guide
